+++ pyproject.toml
@@ -1,21 +1,21 @@
 [tool.poetry]
 name = "servicex"
-version = "2.0.0-alpha.1"
+version = "3.0.0-alpha.2"
 description = ""
 authors = [
     "Ben Galewsky <bengal1@illinois.edu>",
     "Gordon Watts <gwatts@uw.edu>",
 ]
 readme = "README.md"
-packages = [{ include = "servicex_client" }]
+packages = [{ include = "servicex" }]
 
 [tool.poetry.scripts]
-servicex = "servicex_client.app.main:app"
+servicex = "servicex.app.main:app"
 
 [tool.poetry.dependencies]
 python = "^3.8, < 3.12"
 qastle = "^0.16"
-func_adl = "3.2.5"
+func_adl = "^3.2.6"
 requests = "^2.31"
 pydantic = "^1.10"
 httpx = "^0.24"
@@ -33,6 +33,7 @@ fastparquet = { version = "^2023.4.0", optional = true }
 
 sphinx = { version = "^7.0.1", optional = true }
 furo = { version = "^2023.5.20", optional = true }
+jupyter = "^1.0.0"
 
 
 [tool.poetry.group.test.dependencies]
