+++ servicex/app/main.py
@@ -25,8 +25,49 @@
 # CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-from typing import Union
+from typing import Optional
 
-from servicex_client.dataset_identifier import DataSetIdentifier, FileListDataset
+import rich
+import typer
+from servicex._version import __version__
+from servicex.app.cli_options import url_cli_option, backend_cli_option
+from servicex.app.transforms import transforms_app
+from servicex.app.cache import cache_app
+from servicex.servicex_client import ServiceXClient
 
-DID = Union[DataSetIdentifier, FileListDataset]
+app = typer.Typer(no_args_is_help=True)
+
+app.add_typer(transforms_app)
+app.add_typer(cache_app)
+
+
+def show_version(show: bool):
+    """Display the installed version and quit."""
+    if show:
+        rich.print(f"garden-ai {__version__}")
+        raise typer.Exit()
+
+
+@app.callback()
+def main_info(
+    version: Optional[bool] = typer.Option(
+        None, "--version", callback=show_version, is_eager=True
+    )
+):
+    """
+    🌱 Hello, Garden 🌱
+    """
+    pass
+
+
+@app.command(no_args_is_help=False)
+def codegens(
+        url: Optional[str] = url_cli_option,
+        backend: Optional[str] = backend_cli_option
+):
+    sx = ServiceXClient(url=url, backend=backend)
+    rich.print_json(data=sx.get_code_generators())
+
+
+if __name__ == "__main__":
+    app()
