+++ servicex/dataset_group.py
@@ -25,49 +25,30 @@
 # CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-from typing import Optional
+import asyncio
+from typing import Coroutine, List
 
-import rich
-import typer
-from servicex_client._version import __version__
-from servicex_client.app.cli_options import url_cli_option, backend_cli_option
-from servicex_client.app.transforms import transforms_app
-from servicex_client.app.cache import cache_app
-from servicex_client.servicex_client import ServiceXClient
-
-app = typer.Typer(no_args_is_help=True)
-
-app.add_typer(transforms_app)
-app.add_typer(cache_app)
+from servicex.models import TransformedResults
+from make_it_sync import make_sync
 
 
-def show_version(show: bool):
-    """Display the installed version and quit."""
-    if show:
-        rich.print(f"garden-ai {__version__}")
-        raise typer.Exit()
+class DatasetGroup:
+    def __init__(self, datasets: List[Coroutine]):
+        r"""
+        Simultaneously submit a group of transform requests and offer the ability to wait
+        for them all to finish.
 
+        :param datasets: List of transform requests form dataset instances
+        """
+        self.datasets = datasets
 
-@app.callback()
-def main_info(
-    version: Optional[bool] = typer.Option(
-        None, "--version", callback=show_version, is_eager=True
-    )
-):
-    """
-    🌱 Hello, Garden 🌱
-    """
-    pass
+    async def gather_results_async(self) -> List[TransformedResults]:
+        r"""
+        Await on all of the transform requests in the group.
 
+        :return: List of TransformedResults instances. Use the title property to relate the
+                 result to your submission
+        """
+        return await asyncio.gather(*self.datasets)
 
-@app.command(no_args_is_help=False)
-def codegens(
-        url: Optional[str] = url_cli_option,
-        backend: Optional[str] = backend_cli_option
-):
-    sx = ServiceXClient(url=url, backend=backend)
-    rich.print_json(data=sx.get_code_generators())
-
-
-if __name__ == "__main__":
-    app()
+    gather_results = make_sync(gather_results_async)
