+++ servicex/func_adl/util.py
