+++ tests/test_dataset_identifier.py
@@ -25,7 +25,7 @@
 # CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-from servicex_client.dataset_identifier import DataSetIdentifier, RucioDatasetIdentifier, \
+from servicex.dataset_identifier import DataSetIdentifier, RucioDatasetIdentifier, \
     FileListDataset
